# backend/docs/swagger.yaml
basePath: /
definitions:
...
  trip.ActivityItem:
    properties:
      currency:
        example: KZT
        type: string
      day_label:
        example: Day 1
        type: string
      description:
        example: Cultural first-day itinerary stop
        type: string
      id:
        example: f42f0f73-f08d-4ca6-a74f-e15a028ef70e
        type: string
      kind:
        example: place
        type: string
      location:
        example: Tokyo
        type: string
      manually_added:
        example: false
        type: boolean
      price:
        example: 16000
        type: integer
      source_link:
        example: https://maps.google.com/?q=Sensoji
        type: string
      source_name:
        example: Google Maps
        type: string
      title:
        example: Asakusa and Senso-ji
        type: string
    type: object
  trip.BudgetSummary:
    properties:
      bonus_amount:
        example: 18580
        type: integer
      cashback_amount:
        example: 46450
        type: integer
      currency:
        example: KZT
        type: string
      estimated_food_total:
        example: 84000
        type: integer
      estimated_local_transport_total:
        example: 36000
        type: integer
      events_total:
        example: 37000
        type: integer
      grand_total:
        example: 929000
        type: integer
      halyk_offer_label:
        example: Pay fully with Halyk mock card and unlock cashback
        type: string
      hotel_total:
        example: 355000
        type: integer
      insurance_estimate:
        example: 22000
        type: integer
      transport_total:
        example: 395000
        type: integer
    type: object
...
  trip.ChatMessageDTO:
    properties:
      action:
        example: collect_fields
        type: string
      content:
        example: Change the hotel to a more budget-friendly option near Shinjuku.
        minLength: 2
        type: string
    required:
    - content
    type: object
  trip.ChatResponse:
    properties:
      assistant_hints:
        example:
        - Weather and seasonality are shown as guidance
        - ' not real-time facts.'
        - AI suggestions should be checked again closer to departure.
        items:
          type: string
        type: array
      messages:
        items:
          $ref: '#/definitions/trip.ChatMessage'
        type: array
      missing_fields:
        example:
        - start_date
        - end_date
        items:
          type: string
        type: array
      session_id:
        example: fbc99327-dcb9-4d3e-9367-29c618a5904c
        type: string
    type: object
  trip.CreateTripDTO:
    properties:
      title:
        example: Family trip to Japan in July with budget 900000 from Almaty and Kazakhstan
          passport
        minLength: 3
        type: string
    required:
    - title
    type: object
  trip.HotelOption:
    properties:
      currency:
        example: KZT
        type: string
      description:
        example: Central stay for family and solo travelers
        type: string
      id:
        example: 314ffcbe-5599-4865-aa43-95e5f6f9845a
        type: string
      location:
        example: Shinjuku
        type: string
      name:
        example: Tokyo Family Smart Hotel
        type: string
      price:
        example: 355000
        type: integer
      provider:
        example: Booking Mock
        type: string
      rating:
        example: 4.8
        type: number
      review_link:
        example: https://maps.google.com/?q=Shinjuku+hotel
        type: string
      selected:
        example: true
        type: boolean
    type: object
  trip.ManualActivityDTO:
    properties:
      day_label:
        example: Day 3
        type: string
      description:
        example: Manual event insertion from the selection screen
        type: string
      kind:
        example: event
        type: string
      location:
        example: Tokyo
        type: string
      price:
        example: 21000
        type: integer
      source_link:
        example: https://kino.kz
        type: string
      source_name:
        example: Kino.kz
        type: string
      title:
        example: Kino.kz Anime Event Pick
        type: string
    required:
    - day_label
    - kind
    - location
    - title
    type: object
  trip.OfferSummary:
    properties:
      bonus_amount:
        example: 18580
        type: integer
      cashback_amount:
        example: 46450
        type: integer
      halyk_offer:
        example: Pay fully with Halyk mock card and unlock cashback
        type: string
      highlights:
        example:
        - Mock cashback applied on full Halyk payment
        - Bonus estimate included for pitch and UI
        - Kino.kz suggestions included where relevant
        items:
          type: string
        type: array
    type: object
  trip.PatchTripDTO:
    properties:
      budget:
        example: 900000
        type: integer
      citizenship:
        example: Kazakhstan
        type: string
      destination_city:
        example: Tokyo
        type: string
      destination_country:
        example: Japan
        type: string
      end_date:
        example: "2026-07-17"
        type: string
      event_interest:
        example: true
        type: boolean
      hotel_preferences:
        example:
        - family-friendly
        - city-center
        items:
          type: string
        type: array
      insurance_needed:
        example: true
        type: boolean
      interests:
        example:
        - culture
        - food
        - events
        items:
          type: string
        type: array
      origin_city:
        example: Almaty
        type: string
      start_date:
        example: "2026-07-10"
        type: string
      transport_type:
        example: flight
        type: string
      travelers:
        items:
          $ref: '#/definitions/trip.TravelerInput'
        type: array
      trip_purpose:
        example: family
        type: string
    type: object
  trip.ReviewSummary:
    properties:
      id:
        example: 9c0b3bb5-42fc-4ae8-8db2-1aa02d9a20e0
        type: string
      kind:
        example: hotel
        type: string
      source_link:
        example: https://tripadvisor.com
        type: string
      source_name:
        example: Tripadvisor
        type: string
      summary:
        example: Review summary highlights cleanliness, transit access, and family
          room comfort.
        type: string
      target_name:
        example: Tokyo Family Smart Hotel
        type: string
    type: object
...
  trip.TodoSection:
    properties:
      id:
        example: activities
        type: string
      items:
        items:
          $ref: '#/definitions/trip.TodoItem'
        type: array
      title:
        example: Places and Events
        type: string
    type: object
  trip.TransportOption:
    properties:
      arrival:
        example: "16:20"
        type: string
      currency:
        example: KZT
        type: string
      departure:
        example: "07:10"
        type: string
      description:
        example: High-comfort long-haul option
        type: string
      destination:
        example: Tokyo
        type: string
      id:
        example: 19798767-f9b0-4ca8-8fe8-5d51706caf92
        type: string
      mode:
        example: flight
        type: string
      origin:
        example: Almaty
        type: string
      price:
        example: 395000
        type: integer
      provider:
        example: JAL Mock
        type: string
      selected:
        example: true
        type: boolean
      title:
        example: Almaty -> Tokyo
        type: string
    type: object
  trip.Traveler:
    properties:
      age_group:
        example: 7-12
        type: string
      id:
        example: 54f0dd78-8c07-4ee6-bc59-e0bd94622b0f
        type: string
      name:
        example: Aru
        type: string
      notes:
        example: Needs stroller-friendly routes
        type: string
      preferences:
        example:
        - anime
        - parks
        items:
          type: string
        type: array
      type:
        example: child
        type: string
    type: object
  trip.TravelerInput:
    properties:
      age_group:
        example: 7-12
        type: string
      name:
        example: Aru
        type: string
      notes:
        example: Needs stroller-friendly routes
        type: string
      preferences:
        example:
        - anime
        - parks
        items:
          type: string
        type: array
      type:
        example: child
        type: string
    required:
    - age_group
    - type
    type: object
...
  trip.TripDetailsResponse:
    properties:
      activities:
        items:
          $ref: '#/definitions/trip.ActivityItem'
        type: array
      budget:
        $ref: '#/definitions/trip.BudgetSummary'
      chat_entrypoints:
        example:
        - Ask AI to refine the plan
        - Ask AI about weather and season
        - Ask AI why this hotel fits
        items:
          type: string
        type: array
      offers:
        $ref: '#/definitions/trip.OfferSummary'
      review_summaries:
        items:
          $ref: '#/definitions/trip.ReviewSummary'
        type: array
      selected_hotel:
        $ref: '#/definitions/trip.HotelOption'
      selected_transport:
        $ref: '#/definitions/trip.TransportOption'
      todo_sections:
        items:
          $ref: '#/definitions/trip.TodoSection'
        type: array
      travelers:
        items:
          $ref: '#/definitions/trip.Traveler'
        type: array
      trip:
        $ref: '#/definitions/trip.Trip'
      visa:
        $ref: '#/definitions/trip.VisaInfo'
    type: object
  trip.VisaInfo:
    properties:
      checklist:
        example:
        - Passport
        - Application form
        - Hotel proof
        - Trip itinerary
        - Insurance
        items:
          type: string
        type: array
      country:
        example: Japan
        type: string
      notes:
        example: Mock data for the hackathon prototype
        type: string
      recommended_lead:
        example: Begin visa preparation 30 days before departure
        type: string
      requirement:
        example: Prototype shows a visa-assistant style checklist
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@yourdomain.com
    name: API Support
    url: http://yourdomain.com/support
  description: 'JWT Token as Bearer: Authorization: Bearer {token}'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://yourdomain.com/terms/
  title: Smart Travel Halyk Backend API
  version: "1.0"
paths:
...
  /api/v1/trips:
    post:
      consumes:
      - application/json
      description: Creates a new trip draft and initializes chat session state.
      parameters:
      - description: Trip draft payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/trip.CreateTripDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/trip.TripDetailsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create trip draft
      tags:
      - '@trip'
  /api/v1/trips/{tripId}:
    get:
      description: Returns the aggregate trip response for dashboard rendering.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.TripDetailsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get trip details
      tags:
      - '@trip'
    patch:
      consumes:
      - application/json
      description: Updates trip-level fields collected before confirmation and generation.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      - description: Trip patch payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/trip.PatchTripDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.TripDetailsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update trip draft fields
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/activities:
    post:
      consumes:
      - application/json
      description: Adds a place or event item to the trip and updates dashboard state.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      - description: Manual activity payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/trip.ManualActivityDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.TripDetailsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Add activity manually
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/budget:
    get:
      description: Returns aggregate trip costs, cashback, bonus, and offer fields.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.BudgetSummary'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get budget summary
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/chat/messages:
    get:
      description: Returns the current chat history attached to a trip planning session.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.ChatResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get trip chat history
      tags:
      - '@trip'
    post:
      consumes:
      - application/json
      description: Adds a user message to the trip planning session and returns AI
        chat output.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      - description: Chat message payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/trip.ChatMessageDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.ChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Send planning chat message
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/confirm:
    post:
      description: Confirms collected fields and generates the master-plan dashboard
        response.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.TripDetailsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Confirm trip and generate plan
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/options/hotels:
    get:
      description: Returns selected and alternative hotel options for a trip.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/trip.HotelOption'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get hotel options
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/options/hotels/{optionId}/select:
    post:
      description: Replaces the currently selected hotel option and recalculates budget.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      - description: Hotel option ID
        in: path
        name: optionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.TripDetailsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Select hotel option
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/options/transport:
    get:
      description: Returns selected and alternative transport options for a trip.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/trip.TransportOption'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get transport options
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/options/transport/{optionId}/select:
    post:
      description: Replaces the currently selected transport option and recalculates
        budget.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      - description: Transport option ID
        in: path
        name: optionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.TripDetailsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Select transport option
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/regenerate:
    post:
      description: Regenerates trip details from current stored state and chat context.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.TripDetailsResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Regenerate trip plan
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/reviews:
    get:
      description: Returns AI-produced review summaries and source links for hotels,
        places, and events.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/trip.ReviewSummary'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get review summaries
      tags:
      - '@trip'
  /api/v1/trips/{tripId}/visa:
    get:
      description: Returns visa assistant information for the selected destination.
      parameters:
      - description: Trip ID
        in: path
        name: tripId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/trip.VisaInfo'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get visa info
      tags:
      - '@trip'
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"